+++ History.md
@@ -1,3 +1,4 @@
+- is maven build profile to gen xslt, mvn clean package -P generateXSLTFromSchematron
 - OXPullprovider to no longer generate invsalid XML if a duedate is set
 
 2.5.1