+++ History.md
@@ -1,6 +1,12 @@
+2.5.2
+=======
+2022-07-09
+
+- Support validation of XRechnung (CII) 2.2 
 - allow to create fx 1 files with command line again
 - is maven build profile to gen xslt, mvn clean package -P generateXSLTFromSchematron
 - OXPullprovider to no longer generate invsalid XML if a duedate is set
+- Add missing encodeXML to node payment terms description #278 thanks to weclapp-dev
 
 2.5.1
 =======